% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlled_rounding_off_5.R
\name{controlled_rounding_off_5}
\alias{controlled_rounding_off_5}
\title{Controlled rounding off numbers}
\usage{
controlled_rounding_off_5(a_vector)
}
\arguments{
\item{a_vector}{a vector of positive/negative numbers}
}
\description{
Return controlled rounding off numbers
}
\examples{
x <- rnorm(10, mean = 10, sd = 10)
y <- controlled_rounding_off_5(a_vector = x)
(c(sum(x), sum(y)))
(z <- cbind(x, y))
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
