% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_counts_2_props.R
\name{convert_2_prop}
\alias{convert_2_prop}
\title{convert a vector of counts into proportions}
\usage{
convert_2_prop(x)
}
\arguments{
\item{x}{a vector (of counts)}
}
\description{
convert a vector of counts into proportions
}
\examples{
x <- 1:5
p <- convert_2_prop(x)
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{counts;}
\keyword{proportions}
