% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jk_replicate_wgts.R
\name{DAGJK_replicate_wgts}
\alias{DAGJK_replicate_wgts}
\title{create DAGJK  replicate weights}
\usage{
DAGJK_replicate_wgts(df, replicate_NO = 100, random_nbr_seed = 1234567)
}
\arguments{
\item{df}{a dataframe, which contains unit_ID, psu_code, stratum_code and selection_wgt--must be in this ordr}

\item{replicate_NO}{number of replicate weights}

\item{random_nbr_seed}{random number seed}
}
\description{
It returns DAGJK (Delete A Group Jackknife) replicate weights
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
