% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_non_unique_rows.R
\name{get_non_unique_rows}
\alias{get_non_unique_rows}
\title{Get non-unique rows in a dataframe}
\usage{
get_non_unique_rows(a_df, a_var)
}
\arguments{
\item{a_df}{a dataframe}

\item{a_var}{a variable, for defining row uniqueness}
}
\description{
Return a dataframe where the the rows are non-unique in terms of a varialbe
}
\examples{
fk_data <-
data.frame(x = c(rep("a", 2), rep("b", 3), rep("c", 5), letters[11:20]),
y = rnorm(20))
test_df <- get_non_unique_rows(a_df = fk_data, a_var = "x")

}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
