% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jk_variance_4_cat.R
\name{jk_variance_4_cat}
\alias{jk_variance_4_cat}
\title{calculate Jackknife variance for a categorical variable}
\usage{
jk_variance_4_cat(df, count = TRUE)
}
\arguments{
\item{df}{a dataframe, which contains a categorical variable, final_wgt and replicate weights--must be in this order}

\item{count}{a logical variable}
}
\description{
It returns total of each category and corresponding variance if count is TRUE; otherwise returns proportion of each category and corresponding variance
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{Jackknife;}
\keyword{deviation}
\keyword{standard}
\keyword{variance;}
