% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_Register.R
\name{my_Register}
\alias{my_Register}
\title{Get all R files under a folder}
\usage{
my_Register(folder_name = "R")
}
\arguments{
\item{folder_name}{name of the folder}
}
\description{
This returns a dataframe, which has all the R-file information under the specified folder
}
\examples{
df <- my_Register()
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
