% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_allocation.R
\name{proportional_allocation}
\alias{proportional_allocation}
\title{Proportional allocation}
\usage{
proportional_allocation(n, a_vec)
}
\arguments{
\item{n}{sample size}

\item{a_vec}{a vector, which contains sizes of strata}
}
\description{
This is to return proportionally allocated sample sizes
}
\examples{
a <- c(100, 200, 300, 200, 400, 500)
(b <- proportional_allocation(200, a))
(sum(b))

}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
