% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_uneq_prob.R
\name{sample_uneq_prob}
\alias{sample_uneq_prob}
\title{Sampling with unequal selection probabilities}
\usage{
sample_uneq_prob(a_pop, relative_prob_wgt, ssize)
}
\arguments{
\item{a_pop}{a population}

\item{relative_prob_wgt}{relative probability weights}

\item{ssize}{sample size}
}
\description{
This is a wraper of sampling:UPbrewer
}
\examples{
popu <- letters[1:3]
relative_prob <- c(0.5, 0.3, 0.1)
sample_uneq_prob(a_pop = popu, relative_prob_wgt = relative_prob, ssize = 2)
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{probabilities}
\keyword{selection}
\keyword{unequal}
