% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wright_allocation.R
\name{wright_allocation}
\alias{wright_allocation}
\title{Wright allocation}
\usage{
wright_allocation(n, Nh, Sh)
}
\arguments{
\item{n}{the total sample size}

\item{Nh}{size of stratum h}

\item{Sh}{standard deviation of y in stratum h}
}
\description{
Return Wright allocation
}
\examples{
n <- 10
Nh <- c(47, 61, 41)
Sh <- c(10, 6, 4)
(the_results <- wright_allocation(n, Nh, Sh))

}
\references{
Wright, T. (2014). A simple method of exact optimal sample allocation under stratification with any mixed contraint patterns
}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
