% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_cumulative_mean.R
\name{z_cumu_mean}
\alias{z_cumu_mean}
\title{cumulative mean}
\usage{
z_cumu_mean(a_vector)
}
\arguments{
\item{a_vector}{a vector of numbers}
}
\description{
return cumulative mean of a vector
}
\examples{
x <- 1:10
(y <- z_cumu_mean(a_vector = x))

}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{mean}
