% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_cumulative_sd.R
\name{z_cumu_sd}
\alias{z_cumu_sd}
\title{cumulative standard deviation}
\usage{
z_cumu_sd(a_vector, divisor = "n_minus_1")
}
\description{
return cumulative standard deviation of a vector
}
\examples{
x <- 1:10
(y <- z_cumu_sd(a_vector = x, divisor = "n_minus_1"))

}
\author{
Lingyun (Larry) Zhang \email{lyzhang10@gmail.com}
}
\keyword{deviation}
\keyword{standard}
